--- Editor/ImporterSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c889f10e8a3cc2f4fb68df3aca88eb77
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: ba83742ac2b2a1a4f85a14386a9b9ac1, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
