--- Editor/ImporterSettingsEditor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6ba11cc6349142f094dd336e8c3aed9a
-timeCreated: 1653436141
