--- Editor/ShaderGraphImporter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3e2e126cee8490444964da55ca1645de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
